Move test/mainnet toggle to common unlock page
