+++ src/components/UnlockWithKeystore.jsx
@@ -1,10 +1,6 @@
 import React, { useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faCircleNotch, faQuestionCircle, faUnlockAlt } from '@fortawesome/free-solid-svg-icons';
-import Switch from 'react-switch';
-import ReactTooltip from 'react-tooltip';
-import colors from 'utils/colors';
-import { NETWORK_REF_MAINNET, NETWORK_REF_TESTNET } from 'utils/network';
+import { faCircleNotch, faUnlockAlt } from '@fortawesome/free-solid-svg-icons';
 import {
   ERROR_FAILED_READING_FILE,
   ERROR_INVALID_KEYSTORE,
@@ -15,11 +11,9 @@ import { wait } from 'utils/wait';
 import Alert, { ALERT_TYPE_INFO } from 'components/Alert';
 import Button from 'components/Button';
 import { ErrorMessage, Input, InputGroup, Label } from 'components/Forms';
-import { useIconService } from 'components/IconService';
 import { useWallet } from 'components/Wallet';
 
 function UnlockWithKeystore({ onUnlockWallet }) {
-  const { changeNetwork, network } = useIconService();
   const { unlockWallet } = useWallet();
   const passwordInput = useTextInput('');
   const [keystoreFile, setKeystoreFile] = useState(null);
@@ -69,10 +63,6 @@ function UnlockWithKeystore({ onUnlockWallet }) {
     return !errors.password && !errors.keystoreFile;
   }
 
-  function handleChangeNetwork(checked) {
-    changeNetwork(checked ? NETWORK_REF_MAINNET : NETWORK_REF_TESTNET);
-  }
-
   return (
     <form onSubmit={handleOnSubmit}>
       <p>Choose your keystore file and enter the wallet password to unlock your wallet.</p>
@@ -119,48 +109,15 @@ function UnlockWithKeystore({ onUnlockWallet }) {
           {errors.password && <ErrorMessage>{errors.password}</ErrorMessage>}
         </InputGroup>
 
-        <div className="flex items-center justify-between mt-6">
-          <Button type="submit" disabled={isLoading}>
-            <FontAwesomeIcon
-              icon={isLoading ? faCircleNotch : faUnlockAlt}
-              spin={isLoading}
-              fixedWidth
-              className="mr-1 opacity-75"
-            />
-            Unlock{isLoading ? 'ing' : ''} wallet
-          </Button>
-
-          <label className="text-right">
-            <div className="text-xs">
-              Mainnet
-              <button
-                type="button"
-                data-tip="Switch on to connect to mainnet, off connects to testnet"
-                className="text-gray-500 ml-1"
-              >
-                <FontAwesomeIcon icon={faQuestionCircle} />
-              </button>
-              <ReactTooltip
-                place="top"
-                effect="solid"
-                multiline={true}
-                event="click mouseenter"
-                eventOff="click mouseleave"
-                clickable={true}
-              />
-            </div>
-            <Switch
-              checked={network.ref === NETWORK_REF_MAINNET}
-              onChange={handleChangeNetwork}
-              onColor={colors.green['600']}
-              offColor={colors.gray['400']}
-              height={24}
-              width={45}
-              checkedIcon={false}
-              uncheckedIcon={false}
-            ></Switch>
-          </label>
-        </div>
+        <Button type="submit" disabled={isLoading} className="mt-6">
+          <FontAwesomeIcon
+            icon={isLoading ? faCircleNotch : faUnlockAlt}
+            spin={isLoading}
+            fixedWidth
+            className="mr-1 opacity-75"
+          />
+          Unlock{isLoading ? 'ing' : ''} wallet
+        </Button>
       </fieldset>
     </form>
   );
