+++ src/pages/UnlockWallet.jsx
@@ -1,9 +1,14 @@
 import React, { useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faUsb } from '@fortawesome/free-brands-svg-icons';
-import { faKey } from '@fortawesome/free-solid-svg-icons';
+import { faKey, faQuestionCircle } from '@fortawesome/free-solid-svg-icons';
 import { navigate } from '@reach/router';
 import queryString from 'query-string';
+import Switch from 'react-switch';
+import ReactTooltip from 'react-tooltip';
+import colors from 'utils/colors';
+import { NETWORK_REF_MAINNET, NETWORK_REF_TESTNET } from 'utils/network';
+import { useIconService } from 'components/IconService';
 import Layout from 'components/Layout';
 import UnlockWithKeystore from 'components/UnlockWithKeystore';
 import UnlockWithLedger from 'components/UnlockWithLedger';
@@ -27,6 +32,7 @@ function TabButton({ className, isActive, ...props }) {
 }
 
 function UnlockWalletPage({ location }) {
+  const { changeNetwork, network } = useIconService();
   const [unlockMethod, setUnlockMethod] = useState(UNLOCK_METHODS.KEYSTORE);
 
   function onUnlockWallet() {
@@ -34,18 +40,56 @@ function UnlockWalletPage({ location }) {
     navigate(queryParams.redirectTo || '/');
   }
 
+  function handleChangeNetwork(checked) {
+    changeNetwork(checked ? NETWORK_REF_MAINNET : NETWORK_REF_TESTNET);
+  }
+
   return (
     <Layout title="Unlock Existing Wallet">
-      <h2 className="text-2xl uppercase tracking-tight">Unlock existing wallet</h2>
       <div className="sm:flex items-start justify-between">
         <img
           src={authenticationSvg}
           alt="person entering secure website"
-          className="hidden sm:block sm:order-2 sm:w-2/5 max-w-full flex-none sm:ml-6 sm:-mt-6"
+          className="hidden sm:block sm:order-2 sm:w-2/5 max-w-full flex-none sm:ml-6"
         />
 
         <div className="sm:order-1 sm:flex-1">
-          <div className="mt-4">
+          <div className="flex items-start justify-between">
+            <h2 className="text-2xl uppercase tracking-tight">Unlock existing wallet</h2>
+
+            <label className="text-right ml-3 mt-1">
+              <div className="text-xs whitespace-no-wrap">
+                Mainnet
+                <button
+                  type="button"
+                  data-tip="Switch on to connect to mainnet, off connects to testnet"
+                  className="text-gray-500 ml-1"
+                >
+                  <FontAwesomeIcon icon={faQuestionCircle} />
+                </button>
+                <ReactTooltip
+                  place="top"
+                  effect="solid"
+                  multiline={true}
+                  event="click mouseenter"
+                  eventOff="click mouseleave"
+                  clickable={true}
+                />
+              </div>
+              <Switch
+                checked={network.ref === NETWORK_REF_MAINNET}
+                onChange={handleChangeNetwork}
+                onColor={colors.green['600']}
+                offColor={colors.gray['400']}
+                height={24}
+                width={45}
+                checkedIcon={false}
+                uncheckedIcon={false}
+              ></Switch>
+            </label>
+          </div>
+
+          <div className="mt-3">
             <TabButton
               onClick={() => setUnlockMethod(UNLOCK_METHODS.KEYSTORE)}
               isActive={unlockMethod === UNLOCK_METHODS.KEYSTORE}
